pyproject.toml
[project]
name = "mcp-codecartographer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "matplotlib>=3.6",
    "httpx>=0.24",
    "python-dotenv>=1.0.0",
]
